Extract getRole function and reuse on update
