+++ lib/circles/roles/update.js
@@ -1,45 +1,54 @@
 const _ = require("lodash");
+const getRole = require("./getRole");
 
 module.exports = function(req, res) {
+    const cleanRoleParams = cleanParams(req.body, [
+        "title", "effects", "term", "areaOfResponsibility", "desiredCharacteristics"
+    ]);
 
     //TODO: Validate term
-
-    let roleParams = {
-        title: req.body.title,
-        person: req.body.person,
-        effects: req.body.effects,
-        term: req.body.term,
-        areaOfResponsibility: req.body.areaOfResponsibility,
-        desiredCharacteristics: req.body.desiredCharacteristics
-    };
-
-    const cleanRoleParams = _.omitBy(roleParams, _.isUndefined);
-
-    // TODO assert that person exists
 
     if(_.isEmpty(cleanRoleParams)) {
         req.log.warn("Request contained no values to update role with");
         return res.status(400).send({ status: "There are no values in request" });
     }
 
-    req.pgdb.query("SELECT data FROM roles WHERE id=$1", [`${req.circle.circleId}/${req.params.roleId}`], (error, result) => {
-        if(!result.rowCount) {
+    updateRole(req.pgdb, `${req.circle.circleId}/${req.params.roleId}`, cleanRoleParams, (error, role) => {
+        if(error && error.type == "NotFound") {
             return res.status(404).send({
                 status: "Role does not exist"
             });
         }
+        if(error) {
+            req.log.error({ error }, "Error while updating role");
+            return res.sendStatus(500);
+        }
 
-        const updatedRole = _.merge(result.rows[0].data, cleanRoleParams);
-
-        req.pgdb.query("UPDATE roles SET data=$1 WHERE id=$2 RETURNING id, data", [updatedRole, updatedRole.canonicalRoleId], (error, result) => {
-            if(error) {
-                req.log.error({ error }, "Error while updating role");
-                return res.sendStatus(500);
-            }
-
-            const role = result.rows[0].data;
-
-            res.status(200).send({ status: "Role updated", role });
-        });
+        res.status(200).send({ status: "Role updated", role });
     });
 };
+
+function cleanParams(body, params) {
+    let obj = {}
+    params.forEach(param => {
+        if(body[param]) obj[param] = body[param];
+    });
+    return obj;
+}
+
+function updateRole(pgdb, id, params, callback) {
+    getRole(pgdb, id, (error, role) => {
+        if(error) {
+            return callback(error);
+        }
+
+        const updatedRole = _.merge(role, params);
+
+        pgdb.query("UPDATE roles SET data=$1 WHERE id=$2 RETURNING id, data", [updatedRole, updatedRole.canonicalRoleId], (error, result) => {
+            if(error) {
+                return callback(error);
+            }
+            callback(null, result.rows[0].data);
+        });
+    });
+}
